Reuse conversion logic for (float)
We gained some pretty serious code for converting `Value`s to float regs
as part of our math support. Move this to `number.rs` and reuse it for
our `(float)` implementation so it can work on values of unknown types.
